requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.0.21
click==8.2.1
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.37.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==33.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
posthog==5.4.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.0
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
zipp==3.23.0
